xds/testutils: Minor cleanup.
